[BUGFIX] Don't cache detail view
A cached detail view leads to problems if the setting "Logged in FE User" was chosen in FlexForm.
Nevertheless the best way would be to cache the detail view for some reasons (Indexing, Performance, etc...) and split the "Logged in FE User" into another action for an upcoming task.

Related: https://github.com/einpraegsam/femanager/issues/25