[FEATURE] enable admin notification via TypoScript
Currently admin notifications are sent only if the flexform field in the plugin
is filled with an email address.

This patch adds the possibility to enable admin notification also via TypoScript.
When sending the mail, the email address from TypoScript takes precedence over
the adresse from the flexform in case both are filled.