[BUGFIX] Fixes issues with ConfigurationManager injection in case of overridden ConfigurationManagerInterface
In case a 3rd party extension is overriding only the Interface of a class like the
ConfigurationManager it can happen, that we get two instances of ConfigurationManagers.
Flux as an example, only overrides the ConfigurationManagerInterface, not the actual
ConfigurationManager implementation class.
So if this class is used in annotations or ObjectManager->get calls, TYPO3 creates a
second, empty instance of the real extbase ConfigurationManager and not the Overwrite-Class.
This causes multiple potential issues eg. missing ContentObject, etc.