[TASK] Check if given validators from user input is same is set in the TypoScript configuration
This ensures, that it's not possible any more to give any validation string to the eID script for any field. Now it's checked if the given validators for the field are configured in the TypoScript. To handle this, now the tt_content.uid is also passed to the eID script. So there is s check which view is selected in the plugin and that make the decission which validation should be used.
If there is a change in the validation data-attribute (e.g. by browserconsole) or if someone directly wants to check if there is already a field filled in database from any field in fe_users, the message "Error: Field could not be validated" will be returned before any other clientside check will be validated.

Related: https://projekte.in2code.de/issues/32500