[BUGFIX] Fix regression of security update
This commit fixes is a beginner issue. While searching for femanager plugins on a page we sharpen the query to only focus on femanager plugin elements now and not on all content elements. In addition a type cast to string will prevent a possible issue if a femanager plugin is found but there is - for whatever reason - no flexform definition stored on it.
Same type cast for the second public function in PluginRepository.

Related: https://github.com/in2code-de/femanager/issues/120