[BUG] fix email embeddings, fixes #311, #343
Symfony Email class no longer returns the cid:abc123 identifier for
the inline attachment from embedFromPath. Instead we can control the
cid identifier by setting the embed part name.

Another issue was that prepareMailObject replaced the body object of
the Email with its own TextPart object. This is incorrect, as the
body should be a Multipart object that contains the html/text as
well as the embedded images. If we do not setBody ourselves, but
keep to just setting html, the Email object will generate the right
body object itself.

See \Symfony\Component\Mime\Email::generateBody