[FEATURE] Add an option "rather not say" for the gender
It should be possible to have the gender field in the form
while still make it non-required.

Also make the "rather not say" value the default value in order
to avoid users being male by default.

Note: The value 99 for "rather not say" is in line what the
sr_feuser_register and feuserextrafields extensions use.