[BUGFIX] Workaround type confusion in StringUtility::makeEmailArray
This is definitely not the correct place to fix it. This should be fixed in ConfigurationUtility::getValue but that requires a larger refactoring of the entire concept of that Utility, as to why this is a simple band-aid patch to get this working for now.

Related: https://projekte.in2code.de/issues/68082