[TASK] Validate missing captcha in captcha-validated actions
For some reason the CaptchaValidation is
skipped when the parameter is empty. This
causes forms to pass when you just leave
the Captcha input untouched.

This change adds a `validateMissingCaptcha`
method to `AbstractController` which checks
for the presence of the `captcha` argument
in form submissions, coupled with the
extension actually being installed and the
fitting TypoScript config being enabled.

If all of that matches and the captcha
argument is explicitly '' (empty string) we
show the "Wrong captcha code" flash message
and redirect back to the originating action.